+++ src/main.cpp
@@ -8,6 +8,7 @@
 
 #include <Geode/ui/GeodeUI.hpp>
 #include <Geode/ui/Button.hpp>
+#include <Geode/cocos/menu_nodes/CCMenuItem.h>
 
 #include <Geode/modify/MenuLayer.hpp>
 
@@ -15,6 +16,7 @@ using namespace geode::prelude;
 using namespace ads;
 
 class $modify(AdsMenuLayer, MenuLayer) {
+protected:
     bool init() {
         if (!MenuLayer::init()) return false;
 
@@ -28,7 +30,7 @@ class $modify(AdsMenuLayer, MenuLayer) {
                     // log::debug("Token: {}", token);
                 } else {
                     log::warn("Auth failed: {}", res.unwrapErr());
-                    Notification::create("Failed to authorize with Argon", NotificationIcon::Error)->show();
+                    // Notification::create("Failed to authorize with Argon", NotificationIcon::Error)->show();
                 };
             });
 
@@ -49,30 +51,14 @@ class $modify(AdsMenuLayer, MenuLayer) {
 
         // ad button in the bottom menu
         if (auto bottomMenu = this->getChildByID("bottom-menu")) {
-            auto popupButton = Button::createWithNode(
+            auto popupButton = CCMenuItemSpriteExtra::create(
                 CircleButtonSprite::createWithSpriteFrameName(
                     "adIcon.png"_spr,
                     0.875f,
                     CircleBaseColor::Green,
                     CircleBaseSize::MediumAlt),
-                [userId = Mod::get()->getSettingValue<std::string>("user-id")](auto) {
-                    if (userId.empty()) {
-                        createQuickPopup(
-                            "No User ID Set",
-                            "You have not set a User ID yet.\n<cy>Do you want to open the Advertisement Manager and mod settings?</c>",
-                            "No",
-                            "Yes",
-                            [](auto, bool ok) {
-                                if (ok) {
-                                    openSettingsPopup(getMod());
-                                    Notification::create("Opening Advertisement Manager", NotificationIcon::Info)->show();
-                                    web::openLinkInBrowser("https://ads.arcticwoof.xyz/");
-                                };
-                            });
-                    } else {
-                        if (auto popup = AdManager::create()) popup->show();
-                    };
-                });
+                this,
+                menu_selector(AdsMenuLayer::onAdMenuButton));
 
             bottomMenu->addChild(popupButton);
             bottomMenu->updateLayout();
@@ -80,4 +66,24 @@ class $modify(AdsMenuLayer, MenuLayer) {
 
         return true;
     };
+
+    void onAdMenuButton(CCObject* sender) {
+        auto userId = Mod::get()->getSettingValue<std::string>("user-id");
+        if (userId.empty()) {
+            createQuickPopup(
+                "No User ID Set",
+                "You have not set a User ID yet.\n<cy>Do you want to open the Advertisement Manager and mod settings?</c>",
+                "No",
+                "Yes",
+                [](auto, bool ok) {
+                    if (ok) {
+                        openSettingsPopup(getMod());
+                        Notification::create("Opening Advertisement Manager", NotificationIcon::Info)->show();
+                        web::openLinkInBrowser("https://ads.arcticwoof.xyz/");
+                    };
+                });
+        } else {
+            if (auto popup = AdManager::create()) popup->show();
+        };
+    }
 };